Merge develop to master for release 2022.12

See merge request hpctoolkit/hpcviewer!238

+++ edu.rice.cs.hpctree/src/edu/rice/cs/hpctree/internal/config/ScopeTreePainter.java
@@ -112,7 +112,7 @@ public class ScopeTreePainter
 	private static int getZoomFactor() {
 		@SuppressWarnings("restriction")
 		int zoom = DPIUtil.getDeviceZoom();
-		return (int)zoom / 100;
+		return zoom / 100;
 	}
